requirements.txt
fastapi
pydantic_settings
sqlalchemy
alembic
typer
uvicorn[standard]
psycopg2
passlib
python-jose[cryptography]
python-dotenv
rich
pytz
aiohttp
aiosmtpd
requests
httpx
ruff
pillow
aioboto3
python-multipart
pytest
pytest-asyncio
pytest-sugar
psutil
speedtest-cli
numpy==1.*
opencv-python
jinja2
pandas
onnx
onnxruntime
asyncpg